Showing types for found definitions (#13)
Sometimes I want #find to print types of lemmas and definitions which it has found. The reason for that, is that for me, types are much more relevant than names. When I see the list of suggestion, I first try to find the full type which I have in mind, and only then look at the correspondent name.

Here I implement an option find.showTypes, setting which to true will configure #find to show types as well.